+++ CHANGELOG.md
@@ -55,3 +55,6 @@
   - fixed relations update code. Was only updating the number of appearances and not the first/last events.
   - hopefully properly fixed the onstage code, which was silently failing and not erroring out for some reason.
   - continued moving guest notes to guests
+- 2025-04-30:
+  - fixed venue parser not grabbing detail from venue name. This could cause an issue where on insert a venue already exists, but without the split detail it's considered "new" and inserted anyway. Venue parser is still a mess.
+  - added functions to onstage to get a relation/band id for a given url in the onstage tab. If it doesn't exist, it will insert it and return the new id. Before, if a relation or band didn't exist, it would insert NULL. This is fine for bands, but relations shouldn't be NULL.